requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
asyncpg==0.30.0
pandas==2.3.3
numpy==2.2.6